Use merit order to calculate technology load
Loads for "connection" nodes is now determined with the aid of the merit
order. Dispatchable producers are (currently) declared by omitting the
"load" and "profile" attributes in favour of "capacity", which declares
the maxiumum acceptable load for the dispatchable.

Other technologies may either use a "profile" from which the load will
be read, or a "load", in which case the load is assumed to be constant
in each point.