Make sampled curves consistent with resolution change
If the original curve has per-minute values in MWh, we need to alter the
values stored in the sampled curve so that they still represent the same
load over time.

For example, if the original curve has a value of 2 MWh in the first
point (minute), that translates to a total value per hour of 120 MWh.
Therefore the 2 MWh value needs to be multiplied by 60 to represent the
same load over time.