Add "Download Curve" button
Since the entire curve is (currently) available to the browser, we don't
go back to the server for the load curve, but generate a CSV file
on-the-fly.

This won't work if we start sending down-sampled curves to the browser,
or on older browsers which have limits on "data:" href lengths.

It will suffice, for the moment, until we have a more "final" UI in
place.