Create demand- and capacity-scaled profiles
When uploaded, we'll process load profile CSVs to create versions scaled
for use with "demand" and "capacity".

Demand-scaled curves are altered such that the sum of all points is 1.0,
permitting use with an annual demand. Capacity-scaled curves have their
maximum value set to 1.0, and all other values scaled relative to this.

See documentation for Paperclip::ScaledCurve (in lib/paperclip) for
examples.