Support exporting TGs to national-scale scenarios
Adds a new Export class which will take a testing ground and create a
full-size national scenario on ETEngine accounting for the inputs in the
original scenario, and changes made in the testing ground.

Options for importing scenarios from ETEngine on localhost or
etengine.dev have been removed. A global config option specifying the
ETEngine and ETModel paths will be added at a later date.