Add support for local, battery-style storage
Adds basic battery storage functionality, which will store the energy
produced by a node when there is an surplus, and release it when there
is a deficit.

The implementation is described as "local storage" since only energy
produced within the node containing the battery is stored, and only
demand for energy within the node will cause the battery to be
discharged.

Non-local storage, i.e. storing energy produced in other network
components, is not yet supported.

A storage component is defined in the testing ground like any other
technology. Omit the usual "capacity" and "profile" attributes, and
instead use "storage" to define the maximum amount of energy which may
be held in the battery.