Change "point" to "frame"
Inspired by video game programming, "point" has been changed to "frame".
A frame represents a time-period for which we calculate the load of
technologies, and flows throughout the network.

The number of frames calculated depend on the length of the load curves
assigned to technologies. Load curves of length 8760 have one frame
representing each one-hour period in the year. Curves with a length of
35040 indicate the year is split into 15 minute frames.

Occasionally, we also referred to these as "time-steps".

See also: http://gameprogrammingpatterns.com/game-loop.html