Add PreemptiveConsumer
Represents a technology which, when flexibility is enabled, may add load
to the network -- provided there is spare capacity -- instead of adding
it later when the profile specifies.

This is a form of demand-shifting whereby we model a technology being
turned on earlier than the profile specifies in order to minimise the
chances of capacity exceedance later.