DeferrableConsumer should respect own capacity
Fixes deferrable to not exceed its own capacity limit, except in the
final step when there may be multiple unfulfilled deferred loads.

Fixes the calculator so that deferrable does not store the same deferred
load multiple times in the event that the technology received no load in
the first frame.