Prevent strategy picker showing selected strategies
The multi-select plugin would change the name of the strategy picker
button to include the chosen options. However, the names are too long to
be shown, and users may be confused why a drop-down look-a-like has
suddenly changed into something different.

The button will now always display the button title and the number of
currently selected strategies.