Add support for buffer availability and volume
Buffers now require two profiles, a base "availability" profile and a
"use" profile. Availability profiles describe the minimum amount which
should be stored in the buffer at each moment in time, while "use"
describes that which is taken out of the buffer.

If there is insufficient energy stored in the Buffer, the excess use
goes unmet, and is assumed to me made-up by a gas burner (not yet
modelled).

Buffer behaviour is identical when buffering is disabled, save for the
fact that it will not try to pre-charge the buffer when there is spare
network capacity, and will follow the availability profile precisely.

Closes #266
Closes #299
Ref #270
Ref #296