Add all-or-nothing subnet calculation
Calculates the network bottom-up, assigning to each node as much
consumption as it produces until either all the excess is used, or all
consumption has been satisfied. Then moves up one level in the network
and continues. This repeats until it reaches the network head node.

Endpoints must produce enough energy to completely satisfy the demands
of a technology in order for the technology to be assigned a load. If
supply is unsufficient, it must wait until the calculator moves up a
level.

This results in uneven loads even when endpoints have identical
technologies, however the result is still fairer than the non-subnet
calculator.

Further improvements will be forthcoming.