Support decays in Network::Reserve
The given proc will be executed when determining the starting amount in
each frame. The proc is provided the frame number and amount stored in
the previous frame, and may return a positive numeric which should be
subtracted from the reserve.

This may be used to represent decay within the reserve, or to satisfy
consumption of a related technology.