Buffer no longer uses emit/retain
Previously, Buffer would emit the energy it had stored in the buffer,
only to consume it again (as mandatory). This had zero net load on the
network, and was used to tell the buffer how much energy it would start
with in each frame.

Instead, the energy is automatically retained using Reserve, and
consumption is modelled using the reserve's decay function.

This avoids the need to add an explicit "got x mandatory" message when
implementing composite technologies.