Resurrect old Buffer as LegacyBuffer :rage:
Will be used for Buffers which do not belong to a composite. The older
"smart" behaviour is still present.

This was an act-of-last-resort because I can't get the composite-based
Buffer to work properly when not a member of a composite. And converting
the profiles to be capacity based was proving tricky and bug prone.