Fix poor performance of shared buffers
Removes SimpleDelegator, which checks for the existance of a method
prior to calling it, with FastDelegator. The replacement creates the
delegation class at compile time and skips much of the expensive
overhead of its predecessor.

It is less flexible, and more likely to break, but the performance
difference in the calculation loop is substantial. Given a simple
network with a shared buffer and a single attached heat pump:

  Before: 18.691 seconds
  After:   3.589 seconds