P2P emits gas onto the gas network
The lack of multi-carrier support in technologies results in an ugly
solution to the problem. The gas builder assumes that wherever there is
a "siphon" (currently used only for P2P) that there is also a gas
endpoint.

The Siphon takes the path from the gas endpoint to the head gas node and
sends energy back along that path, resulting in gas production.

Ref #926