Remove legacy InstalledTechnology#behavior
The behavior is defined by the Technology, and it is no longer necessary
for the installed tech to customise this. This also triggered #1100
since the behavior would return nil since the addition of YAML-based
technologies. It would trip up the gas network builder in the event
that a Siphon (P2H) was installed with no other gas technologies, since
"behavior" was expected to return "siphon" and not nil.

Closes #1100