VariantKwh returns [] when there is no variant
If the user is trying to measure the kWh of gas or heat, and there is no
gas or heat node for the corresponding electricity endpoint (no
applicable technologies installed), VariantKwh should return an empty
array indicating that there are no loads to be measured.

Previously, 0.0 would be returned which the measure took to mean that
there was a single load of 0.0.

Closes #432.