Add sliders for the heat asset distribution fields
Adds new events to editable table to be able to set up and reconfigure
sliders as necessary.

Adds a customised version of jquery.linkedsliders, made to work with the
Boostrap sliders.

Performance is poor due to EditableTable saving the form values to a
JSON field every time a change occurs.