Refactor CurveComponent; catch invalid byte errors
In order to catch the UTF-8 byte errors prior to Paperclip processing, a
`before_post_process` hook was needed. Unfortunately this was not
available in CurveComponent due to `has_attached_file` being called on
subclasses AFTER the CurveComponent module was included.

CurveComponent is now a module containing a single `module` method which
will instead build a dynamic module for inclusion. This allows us to
call `has_attached_file` in the module and then add the hook.

Closes #1426
Closes #1428